QuorumProxy is heavily used object (as its exposed to all data stores / DDSs) and thus we do see Node warnings in various stress tests about exceeding event listern limit (of 10). Raise this limit to 50 and move implementation to client as it does not belong on the server side.
